requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
